Restores login to working order.
Previous login logic would try to submit both the username and password at once, when they did not appear to be an email address. The latest version of the website expects just one field to be entered and then redirects to a form that contains both, after a short delay. Logic for handling this scenario already existed so we just use that by default. Unfortunately, the new submit button does not work with the click method so we use submit, instead